Refactor util structure to better reflect purpose Revamped print friendly structure to hopefully make it readable by ATSes

+++ src/utils/dateUtils.js
@@ -7,7 +7,7 @@ export const useDateUtils = () => {
         const date = new Date(Date.parse(rawDate));
 
         const options = {
-            month: 'numeric',
+            month: 'short',
             year: 'numeric'
         };
